feat(): route messages through dexie
